Designate Home Route
I designated the home route to be defined at /home instead of it inherently being known as "/"

What will this do. This will alter the nav to have the basis home view of home. However when utilizing the nav it won't load /home/home or /home/about it will simply load it on the root folder!