Get Started (#1)
* Install Vue Router

I installed the Vue Router to the project!

* Add Vuex - State

I added Vuex for state management!

* Designate Home Route

I designated the home route to be defined at /home instead of it inherently being known as "/"

What will this do. This will alter the nav to have the basis home view of home. However when utilizing the nav it won't load /home/home or /home/about it will simply load it on the root folder!

* Modernize routes + Name site

I modernized the routes to have information for metadata. I also gave the home page a name "Larson's Site"